+++ src/types/props.ts
@@ -25,10 +25,10 @@ export type iInput = iExtendedElement<HTMLInputElement> & {
 };
 
 /**
- * Textarea component's type that extends the `<textarea>` element [Read more](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/textarea).
+ * TextArea component's type that extends the `<textarea>` element [Read more](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/textarea).
  */
-export type iTextarea = iExtendedElement<HTMLTextAreaElement> & {
-  refTextarea?: Ref<HTMLTextAreaElement>;
+export type iTextArea = iExtendedElement<HTMLTextAreaElement> & {
+  refTextArea?: Ref<HTMLTextAreaElement>;
   label?: string;
   error?: string;
   maxWidth: boolean;